requirements.txt
python-telegram-bot==20.7
flask==3.0.0
yt-dlp==2023.12.30
requests==2.31.0
